social-app-backend/requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.8.1
anyio==4.12.0
asgiref==3.11.0
async-timeout==5.0.1
asyncpg==0.31.0
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.4
boto3==1.34.10
botocore==1.34.162
CacheControl==0.14.4
cachetools==6.2.4
celery==5.3.4
certifi==2025.11.12
cffi==2.0.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
daphne==4.0.0
distro==1.9.0
dj-database-url==2.1.0
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dotenv==0.9.9
drf-spectacular==0.27.0
exceptiongroup==1.3.1
Faker==22.0.0
fastapi==0.124.4
ffmpeg-python==0.2.0
filelock==3.20.1
firebase-admin==6.3.0
fsspec==2025.12.0
future==1.0.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-firestore==2.22.0
google-cloud-storage==3.7.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.36.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.1
MarkupSafe==3.0.3
mpmath==1.3.0
msgpack==1.1.2
networkx==3.4.2
numpy==2.2.6
openai==1.6.1
packaging==25.0
paypalrestsdk==1.13.1
pillow==10.2.0
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==6.33.2
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==5.0.1
referencing==0.37.0
regex==2025.11.3
requests==2.31.0
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.10.4
safetensors==0.7.0
sentencepiece==0.1.99
sentry-sdk==1.39.2
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.4
starlette==0.50.0
stripe==7.9.0
sympy==1.14.0
tokenizers==0.15.2
tomli==2.3.0
torch==2.1.2
tqdm==4.67.1
transformers==4.36.2
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
websockets==15.0.1
whitenoise==6.6.0
zope.interface==8.1.1
